feat: rename `ActionButtonDropdown` extensible area
In the new UI in 3.1 there is no longer an actions button, it has been
replaced by the media area button. So this commit replaces the
`ActionButtonDropdown` by the `MediaArea` extensible area.
Adds a sample for this new media area.
Updates the sample's gif.
Updates references to `ActionButtonDropdown` extensible area in the docs.